Add files via upload
This project involves creating a dynamic and user-friendly portfolio generation website. Users input their details through a form in the index.html page, and upon submission, the data seamlessly transfers to the portfolio.html page. The portfolio.html page showcases the user's information in sections such as About, Skills, Education, Projects, Resume, and Contact. The styling is handled by a clean and responsive CSS design. The JavaScript functionality ensures a smooth transition of user data and enhances the interactive features, making it a practical tool for individuals, especially in the tech industry, to easily create and manage their professional portfolios.